refactore:moved the fade effect changes to new hook
