+++ app/components/table.tsx
@@ -1,5 +1,6 @@
 import type { TdHTMLAttributes } from "react";
-import React, { useEffect, useRef, useState } from "react";
+import React from "react";
+import { useTableOverFlow } from "~/hooks/use-table-overflow";
 import { tw } from "~/utils/tw";
 
 export function Table({
@@ -9,36 +10,7 @@ export function Table({
   children: React.ReactNode;
   className?: string;
 }) {
-  const containerRef = useRef<HTMLDivElement>(null);
-  const [isOverflowing, setIsOverflowing] = useState(false);
-
-  useEffect(() => {
-    const checkOverflow = () => {
-      const container = containerRef.current;
-      if (container) {
-        const hasOverflow = container.scrollWidth > container.clientWidth;
-        const hasReachedEnd =
-          container.scrollLeft + container.clientWidth >= container.scrollWidth;
-        setIsOverflowing(hasOverflow && !hasReachedEnd);
-      }
-    };
-
-    checkOverflow(); // Initial check
-    window.addEventListener("resize", checkOverflow); // Check on resize
-
-    // Ensure the scroll event listener is added correctly
-    const container = containerRef.current;
-    if (container) {
-      container.addEventListener("scroll", checkOverflow);
-    }
-
-    return () => {
-      window.removeEventListener("resize", checkOverflow);
-      if (container) {
-        container.removeEventListener("scroll", checkOverflow);
-      }
-    };
-  }, []);
+  const {containerRef, isOverflowing} = useTableOverFlow()
 
   return (
     <div className={`relative ${isOverflowing ? "overflowing" : ""}`}>
